% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{analyticForm}
\alias{analyticForm}
\title{Calculates probabilities for power/sensitivity calculations}
\usage{
analyticForm(single_pp, nn.list, passCombinations, INCREASE)
}
\arguments{
\item{nn.list}{array of stage sizes}

\item{passCombinations}{a list of lists with generated combinations for all stages}

\item{pp}{probability to evaluate}
}
\value{
a data object with probabilities and labels
}
\description{
Calculates probabilities for power/sensitivity calculations
}
