% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Predictive_function.r
\name{Bayesian.predictive.futility.reduction.fun}
\alias{Bayesian.predictive.futility.reduction.fun}
\title{Bayesian predictive design}
\usage{
Bayesian.predictive.futility.reduction.fun(n.list, p.target = 0.27,
  cutoff.n.for.greater.p0 = 0.9, predictive.cutoff = 0.05, beta.a = 1,
  beta.b = 1, outcome.tmp = "event", arm.name = "B", p.h1 = 0.17,
  sim.n = 10000, plot.status = F)
}
\arguments{
\item{n.list:}{sample sizes for each stage (vector length = 2)}

\item{p.target:}{response rate in null hypothesis}

\item{cutoff.n.for.greater.p0:}{the threshold for posterior probability. That is, prob(response rate>p.target|data) >cutoff.n.for.greater.p0}

\item{predictive.cutoff:}{the cutoff for the predictive probability}

\item{beta.a:}{beta prior for the 1st parameter}

\item{beta.b:}{beta prior for the 2nd parameter}

\item{outcome.tmp:}{label for the outcome}

\item{arm.name:}{label for the experimental arm}

\item{p.h1:}{response rate in alternative hypothesis}

\item{sim.n:}{number of simulations}

\item{plot.status:}{indictor whether to generate plot}
}
\description{
The function generates report and statistical plan for futility interim analysis by the Bayesian predictive design for single arm early  phase II clinical trial (p1<p0)
}
\examples{
#--- 15 subjects for the 1st and 2nd stages with H0:prob(response rate>0.2|data) >0.95 and the cutoff for the predictive probability=0.05
Bayesian.predictive.futility.reduction.fun(n1=15,n2=15,p.target=0.20,cutoff.n.for.greater.p0=0.95,predictive.cutoff=0.05,beta.a=1,beta.b=1,outcome.tmp='response',arm.name='A')
}
