% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallApp.R
\name{Bayesian_Predictive_App}
\alias{Bayesian_Predictive_App}
\title{Bayesian predictive design in Shinny application}
\usage{
Bayesian_Predictive_App()
}
\description{
This function will run Shiny application for the Bayesian predictive design in single arm early  phase II clinical trial.  The function will automatically generate a statistical plan.
}
