% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_predictive_Shiny_app.R
\name{Bayesian_predictive_Shinny_app}
\alias{Bayesian_predictive_Shinny_app}
\title{Bayesian predictive design in Shinny application}
\usage{
Bayesian_predictive_Shinny_app()
}
\description{
This function will run Shiny application for the Bayesian predictive design in single arm early  phase II clinical trial.  The function will automatically generate a statistical plan.
}
