% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{bbProb}
\alias{bbProb}
\title{Calculate beta bionomial probabilities}
\usage{
bbProb(n1, n2, beta.a, beta.b, INCREASE = TRUE)
}
\arguments{
\item{n1}{n1}

\item{n2}{n2}

\item{beta.a}{beta.a}

\item{beta.b}{beta.b}

\item{INCREASE}{True (default) or false}
}
\value{

}
\description{
Calculate beta bionomial probabilities
}
