% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{calcBetabinProbabilities}
\alias{calcBetabinProbabilities}
\title{Calculate beta binomial probabilities for both increase and reduce scenario}
\usage{
calcBetabinProbabilities(n1, n2, beta.a, beta.b, INCREASE = TRUE)
}
\arguments{
\item{increase}{}
}
\description{
Calculate beta binomial probabilities for both increase and reduce scenario
}
