% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{calcPbb}
\alias{calcPbb}
\title{Calculate beta binomial probabilities for increase scenario}
\usage{
calcPbb(n1, n2, beta.a, beta.b, k, INCREASE = TRUE)
}
\arguments{
\item{n1}{n1}

\item{n2}{n1}

\item{beta.a}{beta.a}

\item{beta.b}{beta.b}

\item{k}{k value}
}
\value{

}
\description{
Calculate beta binomial probabilities for increase scenario
}
