% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{calcPbbForReduction}
\alias{calcPbbForReduction}
\title{Calculate beta binomial probabilities for reduction scenario}
\usage{
calcPbbForReduction(n1, n2, beta.a, beta.b, k)
}
\arguments{
\item{n1}{n1}

\item{n2}{n2}

\item{beta.a}{beta.a}

\item{beta.b}{beta.b}

\item{k}{k value}
}
\value{

}
\description{
Calculate beta binomial probabilities for reduction scenario
}
