% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{combGen}
\alias{combGen}
\title{A function for generating outcome combinations for calculations using analytic form}
\usage{
combGen(stages, currStage, sizes, sizesCum, reqCum, values)
}
\arguments{
\item{stages}{total number of stages to generate combinations for}

\item{currStage}{value for handling depth of recursion}

\item{sizes}{a vector of sample sizes for the stages}

\item{sizesCum}{a vector of cumulative sample sizes for the stages}

\item{reqCum}{a vector with the required cumulative counts for successful trial outcome}

\item{values}{a list for recursively passing on combinations vector for storing in global list}
}
\value{

}
\description{
A function for generating outcome combinations for calculations using analytic form
}
