% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{evaluateInterim}
\alias{evaluateInterim}
\title{Main function for calculating statistics for the interim study design}
\usage{
evaluateInterim(ns, p.target = 0.2, cutoff.n.for.greater.p0 = 0.95,
  predictive.cutoff = 0.05, beta.a = 1, beta.b = 1,
  outcome.tmp = "response", arm.name = "B", p.h1 = 0.4,
  analysis.type = c("Analytical", "Simulation"), sim.n = 10000,
  seed = 42534253)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Main function for calculating statistics for the interim study design
}
\examples{

}
