% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{generateCombinationsRecursively}
\alias{generateCombinationsRecursively}
\title{Creates different combinations of out trial outcomes for a list stage sizes}
\usage{
generateCombinationsRecursively(sizes, nreqCum)
}
\arguments{
\item{sizes}{an array of stage sizes}

\item{nreqCum}{an array with the required positive outcomes for all stages}
}
\value{
void, results are stored in a global variable
}
\description{
Creates different combinations of out trial outcomes for a list stage sizes
}
