% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{lappend}
\alias{lappend}
\title{Utility function to create list of lists}
\usage{
lappend(lst, ...)
}
\arguments{
\item{lst, }{parent  list to add specified vector to}

\item{..., }{vector to be added to list}
}
\value{
parent list with vector added as last element
}
\description{
Utility function to create list of lists
}
\examples{
newList = lappend(list(), c(1,2,3,4,5))
}
