% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{nNeededFromTotal}
\alias{nNeededFromTotal}
\title{Calculates number of positive outcomes from total number of patients that is required for the full study to be successful}
\usage{
nNeededFromTotal(ns, p.target, beta.a, beta.b, cutoff.n.for.greater.p0,
  increase)
}
\arguments{
\item{ns}{array of patients per stage}

\item{p.target}{target probability for successful study}

\item{beta.a}{beta a}

\item{beta.b}{beta b}

\item{cutoff.n.for.greater.p0}{cutoff.n.for.greater.p0}

\item{increase}{boolean for specifying if study evaluation is aimed capture significant increase or decrease. TRUE = increase FALSE = reduce}
}
\value{
numeric value of outcomes that needs to be exceded for positive outcome of study
}
\description{
Calculates number of positive outcomes from total number of patients that is required for the full study to be successful
}
