% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{prob.fun}
\alias{prob.fun}
\title{Calculates probability for a sequence of outcomes fulfilling the criteria for analytic form power/sensitivity calculations}
\usage{
prob.fun(passCombinationCum, stageSamples, p, INCREASE)
}
\arguments{
\item{p}{probability to test sensitivity/power for}

\item{passCombinations}{an array with outcomes for all stages}

\item{totals}{an array with the total number of patients in the study design}
}
\value{
probability for outcome combination
}
\description{
Calculates probability for a sequence of outcomes fulfilling the criteria for analytic form power/sensitivity calculations
}
