% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{sensAnalysis}
\alias{sensAnalysis}
\title{Simulation based sensitivity analysis}
\usage{
sensAnalysis(pp, n1, n2, r1, r)
}
\arguments{
\item{pp}{pp}

\item{n1}{n1}

\item{n2}{n2}

\item{r1}{r1}

\item{r}{r value}
}
\value{

}
\description{
Simulation based sensitivity analysis
}
