% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{simulateInterims}
\alias{simulateInterims}
\title{Run simulations for sensitivity/power calculations}
\usage{
simulateInterims(senrates, sim.n, n.list, n.predictive.cutoff.list,
  n.needed.for.greater.p0, incrMode, seed = 42534253)
}
\arguments{
\item{seed}{}
}
\description{
Run simulations for sensitivity/power calculations
}
