% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{bbProb}
\alias{bbProb}
\title{Calculate beta bionomial probabilities}
\usage{
bbProb(n1, n2, beta.a, beta.b, INCREASE = TRUE)
}
\arguments{
\item{INCREASE}{}
}
\description{
Calculate beta bionomial probabilities
}
