% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{calcBetabinProbabilitiesInPlace}
\alias{calcBetabinProbabilitiesInPlace}
\title{More efficient implementation for calculating beta binomial distributions, overcoming performance drawback with r-style pushing to arrays}
\usage{
calcBetabinProbabilitiesInPlace(n1, n2, beta.a, beta.b)
}
\arguments{
\item{n1}{n1}

\item{n2}{n2}

\item{beta.a}{beta.a}

\item{beta.b}{beta.b value}
}
\value{

}
\description{
More efficient implementation for calculating beta binomial distributions, overcoming performance drawback with r-style pushing to arrays
}
