% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{calcPbb}
\alias{calcPbb}
\title{Calculate beta binomial probabilities for increase scenario}
\usage{
calcPbb(n1, n2, beta.a, beta.b, k, INCREASE = TRUE)
}
\arguments{
\item{k}{}
}
\description{
Calculate beta binomial probabilities for increase scenario
}
