% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianPredictiveFutility.R
\name{simulateInterims}
\alias{simulateInterims}
\title{Run simulations for sensitivity/power calculations}
\usage{
simulateInterims(
  senrates,
  sim.n,
  n.list,
  n.predictive.cutoff.list,
  n.needed.for.greater.p0,
  incrMode,
  seed = 42534253
)
}
\arguments{
\item{senrates}{senrates}

\item{sim.n}{sim.n}

\item{n.list}{n.list}

\item{n.predictive.cutoff.list}{n.predictive.cutoff.list}

\item{n.needed.for.greater.p0}{n.needed.for.greater.p0}

\item{incrMode}{incrMode}

\item{seed}{seed value}
}
\value{

}
\description{
Run simulations for sensitivity/power calculations
}
